formatting and cleaning up code
