--- data/data.md
@@ -1,2 +1,0 @@
-this folder stores the two sqlite gateway files for this project
-
