+++ kard-CLI/.gitignore
@@ -106,10 +106,9 @@ fabric.properties
 
 # idea settings folder
 .idea/misc.xml
+.idea/workspace.xml
 
 # database files
-.data/account.db
-.data/profile.db
 
 # Maven
 target/
