+++ kard-server/.idea/workspace.xml
@@ -5,21 +5,29 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="498f69f9-90d9-41d8-af1e-dca9bba36060" name="Changes" comment="added docs to NameTest">
-      <change beforePath="$PROJECT_DIR$/../kard-CLI/.idea/misc.xml" beforeDir="false" afterPath="$PROJECT_DIR$/../kard-CLI/.idea/misc.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/../kard-CLI/kard-CLI.iml" beforeDir="false" afterPath="$PROJECT_DIR$/../kard-CLI/kard-CLI.iml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/../kard-CLI/src/CommandLineInterface.java" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/../kard-CLI/src/Main.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/../data/data.md" beforeDir="false" />
       <change beforePath="$PROJECT_DIR$/.gitignore" beforeDir="false" afterPath="$PROJECT_DIR$/.gitignore" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/data/data.md" beforeDir="false" />
       <change beforePath="$PROJECT_DIR$/src/main/Main.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/Main.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/main/controller/AccountController.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/controller/AccountController.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/main/controller/ProfileController.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/controller/ProfileController.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/controller/SpringServerController.java" beforeDir="false" />
       <change beforePath="$PROJECT_DIR$/src/main/controller/StartController.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/controller/StartController.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main/response/ShortResponse.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/response/ShortResponse.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/entity/accounts/Account.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/entity/accounts/Account.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/entity/datafiles/Phone.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/entity/datafiles/Phone.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/entity/profiles/Business.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/entity/profiles/Business.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/entity/profiles/Person.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/entity/profiles/Person.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/request/ContactRequest.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/request/PersonalProfileRequest.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/request/StartRequest.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/response/ResponseContainer.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/response/ShortResponse.java" beforeDir="false" />
       <change beforePath="$PROJECT_DIR$/src/main/usecase/AccountUseCases.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/usecase/AccountUseCases.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/usecase/LoginAuth.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/usecase/LoginAuth.java" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/main/usecase/ProfileUseCases.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/usecase/ProfileUseCases.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/usecase/UUIDGenerator.java" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/util/crypto/CryptoHash.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/util/crypto/CryptoHash.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/util/crypto/SHA256Hash.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/util/crypto/SHA256Hash.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/test/usecase/UUIDGeneratorTest.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/test/usecase/UUIDGeneratorTest.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -52,6 +60,18 @@
       <filtered-out-file-type name="TAG" />
       <filtered-out-file-type name="COMMIT_BY_MESSAGE" />
     </file-type-list>
+  </component>
+  <component name="HighlightingSettingsPerFile">
+    <setting file="file://$PROJECT_DIR$/src/main/database/SQLite/SQLiteDataBaseFile.java" root0="FORCE_HIGHLIGHTING" />
+    <setting file="file://$PROJECT_DIR$/src/main/database/SQLite/SQLiteDataBase.java" root0="FORCE_HIGHLIGHTING" />
+    <setting file="file://$PROJECT_DIR$/src/main/controller/AccountController.java" root0="FORCE_HIGHLIGHTING" />
+    <setting file="file://$PROJECT_DIR$/src/main/entity/accounts/Account.java" root0="FORCE_HIGHLIGHTING" />
+    <setting file="file://$PROJECT_DIR$/src/main/entity/accounts/PersonalAccount.java" root0="FORCE_HIGHLIGHTING" />
+    <setting file="file://$PROJECT_DIR$/src/main/database/gateway/AccountGateway.java" root0="FORCE_HIGHLIGHTING" />
+    <setting file="file://$PROJECT_DIR$/src/main/entity/accounts/CorporateAccount.java" root0="FORCE_HIGHLIGHTING" />
+    <setting file="file://$PROJECT_DIR$/src/main/database/gateway/ProfileGateway.java" root0="FORCE_HIGHLIGHTING" />
+    <setting file="file://$PROJECT_DIR$/src/main/usecase/AccountUseCases.java" root0="FORCE_HIGHLIGHTING" />
+    <setting file="file://$PROJECT_DIR$/src/main/Main.java" root0="FORCE_HIGHLIGHTING" />
   </component>
   <component name="ProjectId" id="20NHUKezy27g1r9RLHqgJiHLBNO" />
   <component name="ProjectLevelVcsManager" settingsEditedManually="true" />
@@ -95,6 +115,7 @@
       <recent name="" />
     </key>
     <key name="MoveClassesOrPackagesDialog.RECENTS_KEY">
+      <recent name="util" />
       <recent name="" />
       <recent name="entity.profiles" />
       <recent name="usecase" />
@@ -214,6 +235,7 @@
       <workItem from="1637728426305" duration="2670000" />
       <workItem from="1637734037485" duration="3149000" />
       <workItem from="1637739969936" duration="18000" />
+      <workItem from="1637798085477" duration="46000" />
     </task>
     <task id="LOCAL-00001" summary="fixed error in Name and started NameTest">
       <created>1636166152716</created>
