+++ kard-server/src/main/entity/datafiles/Phone.java
@@ -33,5 +33,4 @@ public class Phone implements Serializable {
     public int hashCode() {
         return Objects.hash(phone);
     }
-
 }
