+++ kard-server/src/main/entity/profiles/Person.java
@@ -13,13 +13,13 @@ public class Person implements ProfileType, Serializable {
     /** Stores all values as subclasses
      * Getters return all values as Strings
      */
-
+    @Serial
+    private static final long serialVersionUID = 6529685098267757690L;
     private final Name name;
     private final Phone phone;
     private final Email email;
     private final String username;
-    @Serial
-    private static final long serialVersionUID = 6529685098267757690L;
+
 
     public Person(Name name, Phone phone, Email email, String username) {
         this.name = name;
