+++ kard-server/src/main/post/ContactRequest.java
@@ -1,12 +1,12 @@
-package request;
+package post;
 
 /**
  * Class which some JSON objects are serialized into from HTTP POST or GET requests. This one in particular contains
- * the account's username and password for login or signup.
+ * the account's username and the contact's username.
  */
-public class StartRequest {
+public class ContactRequest {
     public String accountUsername;
-    public String accountPassword; // should be hashed already
+    public String contactUsername;
 
     /**
      * Getter for the account username
@@ -17,11 +17,10 @@ public class StartRequest {
     }
 
     /**
-     * Getter for the account password
-     * @return the account password
+     * Getter for the contact username
+     * @return the contact username
      */
-    public String getAccountPassword() {
-        return accountPassword;
+    public String getContactUsername() {
+        return contactUsername;
     }
-
 }
