--- kard-server/src/main/response/ResponseContainer.java
@@ -1,7 +1,0 @@
-package response;
-
-/**
- * Interface which response objects implement
- */
-public interface ResponseContainer {
-}
