+++ kard-server/src/test/usecase/UUIDGeneratorTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import util.UUIDGenerator;
 
 import static org.junit.jupiter.api.Assertions.*;
